Pre-Claw Sequence Commit
Commiting inorder to work on claw sequences.

+++ src/org/usfirst/frc/team2485/auto/SequencedItems/MoveClawVertically.java
@@ -7,9 +7,9 @@ import org.usfirst.frc.team2485.subsystems.Claw;
 public class MoveClawVertically implements SequencedItem {
 	
 	private boolean finished; 
-	private final int setpoint; 
+	private final double setpoint; 
 	
-	public MoveClawVertically(int setpoint) {
+	public MoveClawVertically(double setpoint) {
 		this.setpoint = setpoint; 
 		finished = false; 
 	}
